requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
click>=8.0.0
pytest>=7.0.0
customtkinter>=5.0.0
pillow>=9.0.0
